Prettified index.js, added events

+++ src/commands/core/embed.js
@@ -1,5 +1,5 @@
-const { Client, Message, MessageEmbed, ReactionCollector, User } = require('discord.js');
-const { confirmation } = require('@reconlx/discord.js')
+const { Client, Message, MessageEmbed, ReactionCollector, User, ReactionEmoji } = require('discord.js');
+//const { Console } = require('node:console');
 module.exports = {
     name: 'embedbuilder',
     aliases: ["embed"],
@@ -12,7 +12,7 @@ module.exports = {
     adminOnly: true,
     guildOnly: false,
     permissions: [],
-    clientPermissions: ["VIEW_CHANNEL", "SEND_MESSAGES", "EMBED_LINKS"],
+    clientPermissions: ["VIEW_CHANNEL", "SEND_MESSAGES", "EMBED_LINKS", "MANAGE_MESSAGES"],
     string: [],
     /** 
      * @param {Client} client 
@@ -24,38 +24,57 @@ module.exports = {
         let BaseEmbed = new MessageEmbed()
             .setDescription("\u200b")
         let messageEmbedForEditing = await message.channel.send(BaseEmbed);
-        const msgawait = await message.channel.send("Veuillez patientez pendant l'ajout des réactions.");
-        await Promise.all(['✏️', '💬', '🕵️', '🔻', '🔳', '🕙', '🖼️', '🌐', '🔵', '↩️', '📥', '✅', '📑'].map(x => msgawait.react(x)));
-        await msgawait.edit(":pencil2: Modifier le titre\n:speech_balloon: Modifier la description\n:detective: Modifier l'auteur\n:small_red_triangle_down: Modifier le footer\n:white_square_button: Modifier le thumbnail\n:clock10: Ajouter un timestamp\n:frame_photo: Modifier l'image\n:globe_with_meridians: Modifier l'url\n:blue_circle: Modifier la couleur\n:leftwards_arrow_with_hook: Ajouter un field\n:inbox_tray: Copier un embed existant\n:white_check_mark: Envoyer l'embed\n:bookmark_tabs: Modifier un message du bot avec cet embed");
-
+        const msgawait = await message.channel.send(this.string[0]);
+        await Promise.all(['✏️', '💬', '🕵️', '🔻', '🔳', '🕙', '🖼️', '🌐', '🔵', '↩️', '📥', '✅', '📑', '❌'].map(x => msgawait.react(x)));
+        await msgawait.edit(this.string[1]);
+        /**
+         * 
+         * @param {ReactionEmoji} reaction 
+         * @param {User} user 
+         * @returns The filter of the reactions
+         */
         const filterReaction = (reaction, user) => user.id === message.author.id && !user.bot;
-
+        /**
+         * 
+         * @param {User} m 
+         * @returns The filter of the message
+         */
         const filterMessage = (m) => m.author.id === message.author.id && !m.bot;
 
         const collectorReaction = new ReactionCollector(msgawait, filterReaction);
         collectorReaction.on("collect", async (reaction) => {
-            message.delete();
             switch (reaction._emoji.name) {
                 case "✏️":
                     let msgQuestionTitle = await message.channel.send("Quel est votre titre ?");
-                    const Title = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const Title = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
+                    const noNoWords = [Title];
+
+                    const { content } = message;
+
+                    for (var i = 0; i < noNoWords.length; i++) {
+                        if (content.includes(noNoWords[i])) {
+                            message.delete();
+                        break;
+                        }
+                    }
+                    Title.delete();
                     msgQuestionTitle.delete();
                     BaseEmbed.setTitle(Title);
                     messageEmbedForEditing.edit(BaseEmbed);
-                    Title.delete({ timeout: 5000 });
                 break;
 
                 case "💬":
                     const msgQuestionDescription = await message.channel.send("Quelle est votre description ?");
-                    const Description = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const Description = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
                     msgQuestionDescription.delete();
+                    Description.delete();
                     BaseEmbed.setDescription(Description)
                     messageEmbedForEditing.edit(BaseEmbed);
-                    break;
+                break;
 
                 case "🕵️":
                     const msgQuestionAuthor = await message.channel.send("Quel est le nom votre auteur ?");
-                    const Author = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const Author = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first()
                     await message.channel.send("Voulez-vous ajouter une image à votre auteur ?").then(async (msg) => {
                         msg.delete({ timeout: 10000 });
                         const emoji = await confirmation(msg, message.author, ['✅', '❌'], 10000);
@@ -64,12 +83,14 @@ module.exports = {
                                 m.delete({ timeout: 10000 });
                                 const emoji2 = await confirmation(m, message.author, ['✅', '❌'], 10000);
                                 if (emoji2 === "❌") {
+                                    Author.delete();
                                     BaseEmbed.setAuthor(Author);
                                     messageEmbedForEditing.edit(BaseEmbed);
                                 } else if (emoji2 === "✅") {
-
                                     const msgQuestionUrl = await message.channel.send("Quelle est l'url de votre auteur ?");
-                                    const Url = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                                    const Url = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
+                                    Author.delete();
+                                    Url.delete();
                                     msgQuestionUrl.delete();
                                     BaseEmbed.setAuthor(Author, "", Url)
                                     messageEmbedForEditing.edit(BaseEmbed);
@@ -80,16 +101,20 @@ module.exports = {
                             const AuthorImage = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
                             if (!(AuthorImage.includes("http") || AuthorImage.includes("https") || AuthorImage.includes("png") || AuthorImage.includes("jpg") || AuthorImage.includes("jpeg"))) return message.channel.send("fichiers png, jpg et jpeg seulement !");
                             await message.channel.send("Voulez-vous ajouter une url à votre auteur ?").then(async (mess) => {
+                                mess.delete({ timeout: 10000 });
                                 const emoji3 = await confirmation(mess, message.author, ['✅', '❌'], 10000);
                                 if (emoji3 === "❌") {
+                                    Author.delete();
                                     msgQuestionImage.delete();
                                     BaseEmbed.setAuthor(Author, AuthorImage);
                                     messageEmbedForEditing.edit(BaseEmbed);
                                 } else if (emoji3 === "✅") {
-
                                     const msgQuestionUrl = await message.channel.send("Quelle est l'url de votre auteur ?");
                                     const Url = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
                                     msgQuestionUrl.delete();
+                                    Author.delete();
+                                    Url.delete();
+                                    //AuthorImage.delete();
                                     BaseEmbed.setAuthor(Author, AuthorImage, Url)
                                     messageEmbedForEditing.edit(BaseEmbed);
                                 };
@@ -97,15 +122,16 @@ module.exports = {
                         };
                     });
                     msgQuestionAuthor.delete();
-                    break;
+                break;
 
                 case "🔻":
                     const msgQuestionFooter = await message.channel.send("Quel est votre footer ?");
-                    const Footer = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const Footer = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
                     msgQuestionFooter.delete();
+                    Footer.delete();
                     BaseEmbed.setFooter(Footer)
                     messageEmbedForEditing.edit(BaseEmbed);
-                    break;
+                break;
 
                 case "🔳":
                     const msgQuestionThumbnail = await message.channel.send("Quel est votre thumbnail ?");
@@ -114,7 +140,7 @@ module.exports = {
                     msgQuestionThumbnail.delete();
                     BaseEmbed.setThumbnail(Thumbnail)
                     messageEmbedForEditing.edit(BaseEmbed);
-                    break;
+                break;
 
                 case "🕙":
                     BaseEmbed.setTimestamp()
@@ -123,7 +149,7 @@ module.exports = {
 
                 case "🖼️":
                     const msgQuestionImage = await message.channel.send("Quel est votre image ?");
-                    const Image = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const Image = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content
                     if (!(Image.includes("http") || Image.includes("https") || Image.includes("png") || Image.includes("jpg") || Image.includes("jpeg"))) return message.channel.send("fichiers png, jpg et jpeg seulement !")
                     msgQuestionImage.delete();
                     BaseEmbed.setImage(Image)
@@ -132,42 +158,81 @@ module.exports = {
 
                 case "🌐":
                     const msgQuestionURL = await message.channel.send("Quel est votre url ?");
-                    const url = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const url = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
                     msgQuestionURL.delete();
+                    url.delete();
                     BaseEmbed.setURL(url)
                     messageEmbedForEditing.edit(BaseEmbed);
                 break;
 
                 case "🔵":
                     const msgQuestionColor = await message.channel.send("Quel est votre couleur ?");
-                    const color = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const color = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
+                    color.delete();
                     msgQuestionColor.delete();
                     BaseEmbed.setColor(color)
                     messageEmbedForEditing.edit(BaseEmbed);
-                    break;
+                break;
 
                 case "↩️":
                     const msgQuestionField = await message.channel.send("Quel est votre titre du field ?");
-                    const titleField = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const titleField = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
                     msgQuestionField.delete();
                     const msgQuestionFieldValue = await message.channel.send("Quel est votre value du field ?");
-                    const valueField = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const valueField = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
                     msgQuestionFieldValue.delete();
                     const msgQuestionInline = await message.channel.send("Doit-il être aligné ? (true/false)");
-                    const inlineField = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const inlineField = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
                     msgQuestionInline.delete();
+                    titleField.delete();
+                    valueField.delete();
+                    inlineField.delete();
                     BaseEmbed.addField(titleField, valueField, inlineField)
                     messageEmbedForEditing.edit(BaseEmbed);
-                    break;
+                break;
 
                 case "✅":
                     const msgQuestionEmbed = await message.channel.send("Merci de rentrer l'id du salon où envoyer l'embed");
-                    const channel = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const channel = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first();
                     msgQuestionEmbed.delete();
+                    channel.delete();
                     if (!message.guild.channels.cache.get(channel)) return message.channel.send("Salon invalide");
                     else message.guild.channels.cache.get(channel).send(BaseEmbed);
+                break;
+
+                case "📥":
+                    const msgQuestionEmbedCopy = await message.channel.send("Veuillez envoyer votre object d'embed");
+                    const objectEmbed = (await message.channel.awaitMessages(filterMessage, { max: 1, time: 60000 })).first().content;
+                    const obj = JSON.parse(objectEmbed)
+                    msgQuestionEmbedCopy.delete();
+                    BaseEmbed.edit(obj);
+                    messageEmbedForEditing.edit(BaseEmbed)
+                    //obj.delete();
+                    //message.channel.send({ embed: obj });
+                break;
+
+                case "❌":
+                    msgawait.delete();
+                    messageEmbedForEditing.delete();
                 break;
             }
         });
     },
 };
+
+async function confirmation(message, author, validReactions, time = 60000) {
+    if (!message) throw new ReferenceError('confirmation => "message" is not defined')
+    if (!validReactions || validReactions.length !== 2) throw new ReferenceError('confirmation => Invalid form body [validReactions]')
+    if (typeof time !== "number") throw new SyntaxError('confirmation => typeof "time" must be a number')
+    if (!message.guild.me.hasPermission('MANAGE_MESSAGES')) return console.log(`confirmation err: Discord Client has to have "MANAGE_MESSAGES" permission.`)
+
+    for (const reaction of validReactions) await message.react(reaction)
+
+    const filter = (reaction, user) => validReactions.includes(reaction.emoji.name) && user.id === author.id
+
+    return message
+        .awaitReactions(filter, { max: 1, time: time })
+        .then((collected) => collected.first() && collected.first().emoji.name);
+}
+
+
